Create Print a Message
Objective: Use a while loop to print a message a specified number of times.

Instructions: Ask the user to input a message and the number of times they want it to be printed. Use a while loop to print the message the specified number of times. Increment a counter variable in each iteration to keep track of how many times the message has been printed. Exit the loop when the counter reaches the specified number.